5-th adecvat commit

+++ GameLandWebApplication/Models/GamesUser.cs
@@ -10,7 +10,7 @@ namespace GameLandWebApplication
         public int Id { get; set; }
         public int GameId { get; set; }
         public int UserId { get; set; }
-        public byte Rate { get; set; }
+        public int Rate { get; set; }
 
         public virtual Game Game { get; set; }
         public virtual User User { get; set; }
